добавил .env
