+++ .env
@@ -1,0 +1,2 @@
+interval=1000
+timeout=5000
